requirements.txt
fastapi
uvicorn
python-dotenv
pillow
sqlalchemy
langchain
langchain-groq
python-multipart
google-generativeai